changes in prevision for 32 bits build

+++ srcs/nm_x64.c
@@ -6,7 +6,7 @@
 /*   By: snicolet <snicolet@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/10 03:43:58 by snicolet          #+#    #+#             */
-/*   Updated: 2018/02/13 07:30:30 by snicolet         ###   ########.fr       */
+/*   Updated: 2018/02/13 14:13:28 by snicolet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -54,7 +54,7 @@ static void		handle_x64_list(t_list **lst,
 
 	sym.name = (char*)(size_t)name;
 	sym.type = item->n_type;
-	sym.value = item->n_value;
+	sym.value = (size_t)item->n_value;
 	sym.nsect = (unsigned int)item->n_sect;
 	sym.ndesc = (unsigned int)item->n_desc;
 	ft_lstpush_sort(lst, ft_lstnew(&sym, sizeof(sym)), &handle_sort);
